requirements.txt
Flask
Flask-SQLAlchemy
PyMySQL
cryptography
Flask-Login
Authlib
requests
python-dotenv